refactor(bot): reorganize user edit functionalities #refactor-workspace
- Moved avatar, move, and rename functionalities to separate feature folders
- Updated import paths to reflect new structure

This reorganization improves code readability and maintainability by grouping related functionalities together.
