+++ packages/bot/src/commands/tools/edit.ts
@@ -6,7 +6,6 @@ import type {
 import type { DiscordChannel } from "@dicelette:types/discord";
 import { cmdLn, findln, ln, t } from "@dicelette/localization";
 import type { Translation } from "@dicelette/types";
-import { editUserButtons, selectEditMenu } from "buttons";
 import type { EClient } from "client";
 import { deleteUser } from "database/delete_user";
 import { getDatabaseChar } from "database/get_user";
@@ -15,6 +14,7 @@ import * as Djs from "discord.js";
 import { getEmbeds, getEmbedsList } from "messages/embeds";
 import { findLocation } from "messages/thread";
 import { embedError, haveAccess, reply } from "utils";
+import { editUserButtons, selectEditMenu } from "utils/button";
 
 import { verifyAvatarUrl } from "@dicelette/utils";
