+++ packages/bot/src/features/dice/buttons.ts
