+++ packages/bot/src/features/stats/buttons.ts
@@ -2,10 +2,10 @@ import type { StatisticalTemplate } from "@dicelette/core";
 import { ln } from "@dicelette/localization";
 import type { Settings, Translation } from "@dicelette/types";
 import { isArrayEqual } from "@dicelette/utils";
-import { registerDmgButton } from "buttons";
 import * as Djs from "discord.js";
 import { getEmbeds, parseEmbedFields } from "messages/embeds";
 import { reply } from "messages/send";
+import { registerDmgButton } from "utils/button";
 import { allowEdit } from "utils/check";
 
 /**
